@@ -16,15 +16,15 @@
 		"preview:garden": "vite preview -c 'garden.vite.config.js'"
 	},
 	"devDependencies": {
-		"@gardenjs/render-plugin-svelte": "^1.0.1",
-		"@sveltejs/adapter-static": "^3.0.5",
-		"@sveltejs/kit": "^2.7.1",
-		"@sveltejs/vite-plugin-svelte": "^3.1.2",
-		"express": "^4.21.0",
-		"gardenjs": "^1.3.2",
+		"@gardenjs/render-plugin-svelte": "^2.0.0",
+		"@sveltejs/adapter-static": "^3.0.8",
+		"@sveltejs/kit": "^2.15.2",
+		"@sveltejs/vite-plugin-svelte": "^5.0.3",
+		"express": "^4.21.2",
+		"gardenjs": "^1.4.0",
 		"mdsvex": "^0.12.3",
-		"sass": "^1.79.3",
-		"svelte": "^4.2.19",
-		"vite": "^5.4.8"
+		"sass": "^1.83.1",
+		"svelte": "^5.17.3",
+		"vite": "^6.0.7"
 	}
 }
