@@ -2,11 +2,17 @@
   import { createEventDispatcher } from "svelte";
   import { base } from '$app/paths';
 
-  export let url = "";
   
   const dispatch = createEventDispatcher();
 
-  export let folders = [];
+  /**
+   * @typedef {Object} Props
+   * @property {string} [url]
+   * @property {any} [folders]
+   */
+
+  /** @type {Props} */
+  let { url = "", folders = [] } = $props();
 
   function handleClick() {
     dispatch("toggleMobilenav");
@@ -22,7 +28,7 @@
         <ul>
           {#each folder.items as link}  
             <li>
-              <a href={base}{link.href} class:active={url === base + link.href} on:click={handleClick}>{link.label}</a>
+              <a href={base}{link.href} class:active={url === base + link.href} onclick={handleClick}>{link.label}</a>
             </li>
           {/each}
         </ul>
