@@ -3,9 +3,16 @@
   import Navbar from '../components/header/Navbar.svelte'
   import Vegetables from '../components/vegetables/Vegetables.svelte'
   import Footer from '../components/footer/Footer.svelte'
+  /**
+   * @typedef {Object} Props
+   * @property {import('svelte').Snippet} [children]
+   */
+
+  /** @type {Props} */
+  let { children } = $props();
 </script>
 
 <Navbar showHeroheader />
-<slot></slot>
+{@render children?.()}
 <Vegetables isHome />
 <Footer isHome />
