@@ -1,7 +1,14 @@
 <script>
   import Subpage from '../../layouts/Subpage.svelte'
+  /**
+   * @typedef {Object} Props
+   * @property {import('svelte').Snippet} [children]
+   */
+
+  /** @type {Props} */
+  let { children } = $props();
 </script>
 
 <Subpage>
-  <slot></slot>
+  {@render children?.()}
 </Subpage>
