@@ -1,5 +1,12 @@
 <script>
   import { theme } from "../stores/theme";
+  /**
+   * @typedef {Object} Props
+   * @property {import('svelte').Snippet} [children]
+   */
+
+  /** @type {Props} */
+  let { children } = $props();
 </script>
 
-<slot></slot>
+{@render children?.()}
